feat: Add email template management functionality
- Implemented create, edit, and index views for email templates.
- Removed old campaign and dashboard views.
- Updated routing to reflect new structure.
- Added migration and seeder scripts for database setup.
- Refactored DTOs for better validation and structure.
- Enhanced campaign controller to utilize new template management.

+++ config/database.php
@@ -7,7 +7,7 @@ return [
         'mysql' => [
             'driver' => 'mysql',
             'host' => getenv('DB_HOST') ?: '127.0.0.1',
-            'port' => getenv('DB_PORT') ?: '3306',
+            'port' => (int) getenv('DB_PORT') ?: '3306',
             'database' => getenv('DB_DATABASE') ?: 'mailcamp',
             'username' => getenv('DB_USERNAME') ?: 'root',
             'password' => getenv('DB_PASSWORD') ?: '',
